refactor(stories-brain): rename components/util.tsx to types.ts
