+++ packages/stories/stories-brain/src/components/index.ts
@@ -9,4 +9,4 @@ export * from './FactViewer';
 export * from './PipelinePanel';
 export * from './PredicateList';
 export * from './QueryPanel';
-export * from './util';
+export * from './types';
